Разделить movie_mates/src/main.py на модули #162
