+++ movie_mates/Makefile
@@ -1,0 +1,2 @@
+run_server:
+	python -m src.main
