+++ movie_mates/src/server/events.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class Event(BaseModel):
+    type: str
+    content: str = ''
