+++ movie_mates/src/server/rooms/__init__.py
@@ -1,0 +1,1 @@
+from .rooms import *
